requirements.txt
fastapi
openai
instructor
tqdm
matplotlib
pydot
networkx
instructor
pydantic
typing
python-dotenv
pathlib